// src/enums/brandInformation.ts
export enum BrandInformation {
    REQUIRED_BRAND_NAME = 'O nome da marca é obrigatório.',
    REQUIRED_TEAM_ID = 'O ID da equipe é obrigatório.',
    REQUIRED_USER_ID = 'O ID do usuário é obrigatório.',
    REQUIRED_VALUE_PROPOSITION = 'A proposta de valor é obrigatória.',
    REQUIRED_BRAND_PILLARS = 'Os pilares da marca são obrigatórios.',
    REQUIRED_BRAND_MISSION = 'A missão da marca é obrigatória.',
    REQUIRED_BRAND_INSPIRATION = 'A inspiração da marca é obrigatória.',
    REQUIRED_CURRENT_OBJECTIVE = 'Os objetivos atuais são obrigatórios.',
    REQUIRED_NUMERIC_TARGET = 'A meta numérica é obrigatória.',
    REQUIRED_RESTRICTIONS = 'As restrições são obrigatórias.',
    REQUIRED_BRAND_HASHTAGS = 'As hashtags da marca são obrigatórias.',
    REQUIRED_REFERENCE_CONTENTS = 'Os conteúdos de referência são obrigatórios.',
    REQUIRED_IMPORTANT_DATES = 'As datas importantes são obrigatórias.',
    REQUIRED_RELEVANT_CONTENT = 'Os conteúdos relevantes são obrigatórios.',
    REQUIRED_BRAND_CRISIS = 'O plano de crise da marca é obrigatório.',
    BRAND_ALREADY_EXISTS = 'Já existe uma marca com esse nome.',
    BRAND_ERROR_CREATE = 'Erro ao criar a marca: ',
    BRAND_DELETED = 'Marca deletada com sucesso.',
    BRAND_NOT_FOUND = 'Marca não encontrada.',
}